ml service/requirements.txt
# ── Core web framework ────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── ML / vision ───────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless>=4.9.0

# ── AWS ───────────────────────────────────────────────────────────
boto3>=1.34.0

# ── HTTP client ───────────────────────────────────────────────────
requests>=2.31.0

# ── PDF generation ────────────────────────────────────────────────
fpdf2>=2.7.9

# ── Config / env ──────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.6.0

# ── Testing ───────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
